Change how `documentHovering` is set; cleanup
Because of strange behaviour concerning the `dragLeave` event on the document element, an incremental counter strategy is used to determine if the window is being hovered over or not.
As well, the `holderHovering` property is now set from the document body's drag event handlers for consistency, performance, and code length purposes.

The end result is a more reliable setting of classes based on drag and drop states.